Mv `average_out_cell_energy`, `group_connected_cells`, `average_out_energy_in_connected_cells` to cis_cell.py
